requirements.txt
# === FastAPI Backend ===
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
faiss-cpu
requests
tavily-python
groq
transformers
torch
Pillow
pyjwt
langchain

# === Flask Frontend ===
flask
